% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{binomial_to_LRR_learner}
\alias{binomial_to_LRR_learner}
\title{Converts a binomial learner that predicts `expit` transformed predictions to a relative risk learner that can be used in \code{libraryRR}}
\usage{
binomial_to_LRR_learner(learner)
}
\arguments{
\item{learner}{A binomial \code{sl3} learner.}
}
\description{
Converts a binomial learner that predicts `expit` transformed predictions to a relative risk learner that can be used in \code{libraryRR}
}
