% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\docType{data}
\name{glmnet_screener}
\alias{glmnet_screener}
\title{An \code{sl3} learner object that screens variables using \code{glmnet}.}
\format{
An object of class \code{Lrnr_pkg_SuperLearner_screener} (inherits from \code{Lrnr_base}, \code{R6}) of length 31.
}
\usage{
glmnet_screener
}
\description{
An \code{sl3} learner object that screens variables using \code{glmnet}.
}
\keyword{datasets}
