% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRR_helpers.R
\name{make_fourier_basis_list}
\alias{make_fourier_basis_list}
\title{Generates a default \codE{basis_list} using the fourier basis.}
\usage{
make_fourier_basis_list(V, size = Inf)
}
\arguments{
\item{V}{A vector or matrix of values of `V`. See the corresponding argument in \code{npRR}.}

\item{The}{maximum size of the list.}
}
\value{
A list of \code{fourier_basis} objects.
}
\description{
Generates a default \codE{basis_list} using the fourier basis.
}
