% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npRR.R
\name{npRR}
\alias{npRR}
\title{Machine learning of the covariate-adjusted conditional relative risk function.}
\usage{
npRR(
  V,
  X = NULL,
  A,
  Y,
  weights = NULL,
  library_RR = "autoML",
  library_A = "autoML",
  library_Y = "autoML",
  library_V = "autoML",
  cv_nuisance = TRUE,
  cv_RR = TRUE,
  basis_list = make_fourier_basis_list(V),
  likelihood = NULL
)
}
\arguments{
\item{V}{A vector or matrix of voariates for which to estimate the adjusted conditional relative risk function.}

\item{X}{A vector or matrix of variables to adjust for/marginalize over when computing the conditional relative risk function. If NULL then the unadjusted conditional relative function is computed using \cpde{V}}

\item{A}{A binary vector representing treatment. A = 1 corresponds with recieving treatment and A = 0 corresponds with no treatment. The relative risk is computing comparing as the quotient of the A=1 risk and A=0 risk.}

\item{Y}{A binary vector representing the outcome variable.}

\item{weights}{A vector of weight values to incorporate in all estimation procedures.}

\item{library_RR}{An \code{sl3} \code{Learner} or \code{Stack}  object representing the machine-learning library for estimating the relative risk function. See the R github package \code{\link[sl3]} at \url{https://github.com/tlverse/sl3/tree/master/R}}

\item{library_A}{An \code{sl3} \code{Learner} or \code{Stack}  object representing the machine-learning library for estimating the distribution of \code{A} nuisance parameter, P(A=1|X, V).}

\item{library_Y}{An \code{sl3} \code{Learner} or \code{Stack}  object representing the machine-learning library for estimating the distribution of \code{Y} nuisance parameter, P(Y=1|A, X, V).}

\item{library_V}{An optional \code{sl3} \code{Learner} or \code{Stack}  object representing the machine-learning library for estimating the sequential regression E[E[Y|A=a, V, X]|X]. This need only be supplied when \code{V} is one-dimensional and one wishes to call \code{\link{npRR_inference}}.}

\item{cv_nuisance}{A boolean for whether or not to use cross-validation/cross-validated predictions for the nuisance parameter estimation. This must be true if \code{library_A} or \code{library_V} are \code{sl3} \code{Stack} objects.
If \code{library_A} or \code{library_V} are already cross-validated (Super)learners then it is computationally beneficial to set this parameter tp \code{FALSE}.}

\item{cv_RR}{A boolean for whether to return the cross-validated-selected predictions for the relative risk obtained by cross-validating \code{library_RR}.}

\item{basis_list}{Internal-use variable. list of sieve basis objects to use to convexify the double-robust RR-risk function. See \code{fourier_basis} object as an example.}

\item{likelihood}{A \code{tmle3} fitted likelihood object (as returned by this function) to use as estimates for the nuisance parameters.}
}
\description{
This function allows one to data-adaptively estimate the conditonal relative risk function using machine learning.
Given two sets of covariates `V` and `X`, a binary treatment `A`, and a binary outcome variable`Y`. This function estimates `E[E[Y=1|A=1, V, X] |V]/ E[E[Y=1|A=0, V, X] |V]`. When `X` is omitted, this reduces to `E[Y=1|A=1, V]/ E[Y=1|A=0, X]``
}
