% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npRR.R
\name{npRR_inference}
\alias{npRR_inference}
\title{Function that provides efficient estimates and inference for a data-adaptively kernel smoothed approximation of the covariate-adjusted conditional relative risk function. This method only works when the argument \code{V} provided to \code{\link{npRR}}is univariate.}
\usage{
npRR_inference(
  output,
  kernel_function = function(x, y, sigma) {     exp(-(y - x)^2/(2 * sigma^2))/sqrt(2 *
    3.14 * sigma^2) },
  npoints = 35,
  points = NULL,
  sigmas_grid = NULL,
  min_samples = 100,
  max_samples = Inf,
  VarOverBias = 3
)
}
\arguments{
\item{output}{An output object returned by \code{\link{npRR}}.}

\item{kernel_function}{The kernel function to use for kernel smoothing. By default, the guassian kernel is smooth.
Must be a function of the form `f(x,y,sigma)` where `x` is a vector of values to evaluate the kernel function at, `y` is the centering value of the kernel, and `sigma` is a real-valued smoothness parameter/bandwidth of the kernel.}

\item{npoints}{The number of points at which to estimate the kernel-smoothed relative risk function (equally spaced on the quantile scale).}

\item{points}{A vector of points at which to estimate the kernel smoothed RR estimate. This will override \code{npoints} if supplied.}

\item{sigmas_grid}{A vector of bandwidth values of the kernel to search through. Lepsky's method is used to data-adaptively select an optimal bandwidth.}

\item{min_samples}{The minimum number of samples to be contained in the support of the kernel. This is used to filter out values of \code{sigma_grid} that lead to very narrow kernels.}

\item{max_samples}{The maximum number of samples to be contained in the support of the kernel. This is used to filter out values of \code{sigma_grid} that lead to very wide kernels.}

\item{VarOverBias}{The ratio of the variance and bias of the kernel-smoothed relative risk estimator at a point relative to the true relative risk function.
High values of \code{VarOverBias} lead to undersmoothed, more variable and less biased estimates of the relative risk function. Under suitable conditions, undersmoothing the kernel can lead to the kernel-smoothed relative risk estimates and inference being correct for the true unsmoothed relative risk function.
We stress that the estimates and inference are for the kernel smoothed relative risk function, and not necessarily for the true relative risk function.}
}
\value{
A list containing:
1. A matrix that contains point-estimates and 95`%` pointwise confidence intervals for the kernel-smoothed relative risk function for a number of points. The bandwidth `sigma` used for the kernel smooth at each point is also given.
2. A targeted/updated (TMLE) estimator for the relative risk evaluated at \code{V}
3. A matrix containing the estimated efficient influence function for the kernel-smoothed relative risk function point evaluations. This can be used to construct simultaneous confidence bands.
}
\description{
Function that provides efficient estimates and inference for a data-adaptively kernel smoothed approximation of the covariate-adjusted conditional relative risk function. This method only works when the argument \code{V} provided to \code{\link{npRR}}is univariate.
}
