% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{SL_learner_to_LRR_learner}
\alias{SL_learner_to_LRR_learner}
\title{Converts a name of a \code{\link[SuperLearner]} learner and additional parameters to an \code{sl3} learner that can be used in that can be used in \code{libraryRR}.}
\usage{
SL_learner_to_LRR_learner(name, ...)
}
\arguments{
\item{name}{A character with the name of a \code{\link[SuperLearner]}learner.}

\item{...}{Additional arguments to be passed to the learner.}
}
\description{
Converts a name of a \code{\link[SuperLearner]} learner and additional parameters to an \code{sl3} learner that can be used in that can be used in \code{libraryRR}.
}
