% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\docType{data}
\name{valid_learners}
\alias{valid_learners}
\title{A vector of names of valid \code{\link[SuperLearner]} learners that can be used in \code{Library_RR} for estimating the relative risk.}
\format{
An object of class \code{character} of length 7.
}
\usage{
valid_learners
}
\description{
A vector of names of valid \code{\link[SuperLearner]} learners that can be used in \code{Library_RR} for estimating the relative risk.
}
\keyword{datasets}
