% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{valid_SL_learners}
\alias{valid_SL_learners}
\title{Prints the names of valid \code{\link[SuperLearner]} learners for relative risk estimation.}
\usage{
valid_SL_learners()
}
\description{
Prints the names of valid \code{\link[SuperLearner]} learners for relative risk estimation.
}
